Add support for the Bataineh algorithm and fix touch device support
Bataineh is a new, experimental, algorithm I have worked on implementing.  There have been a number of challenges with it, but it is very close to accurately representing the results in the research paper.
Mobile touch support has also been improved.  I fixed a lot of issues with the W3School code that makes up the slider.