Refactor - Replace Integral Images with cutting edge algorithm (#12)
Integral Image optimization has been replaced by the Chan algorithm, a cutting edge algorithm that is up to 3x faster while using only a fraction of the memory... and allows for processing large image files! 

The default grayscale algorithm has also changed from Qt to Mean.

Variance is now calculated using Population Variance instead of Sample Variance.  This only caused a 1 pixel difference with 2 algorithms.

ISauvola has been abstracted so that it is very easy to make "Improved" versions of any algorithm in the list, not just Sauvola.

Python bindings have been added along with a new set of demos for Python / PIL, C++ Qt and OpenCV, and JS with NodeJS and Web.

Updating the same memory has was proven safe in UpdateToBinary(...), so no temporary image is being used anymore. 

A BinarizationFactory class was added to help simplify using the library and for exposing new language bindings.