DoxaPy - Build system update
Nanobind is correctly building an abi3 shared library, and our wheels and s-dist are now built off of scikit-build-core. 'setup.py' has been deprecated, which is why it is being replaced.

NOTE: macos-latest is not x86! I will be looking at setting up a special wheel for it, if there is interest.